requirements.txt
Flask
python-dotenv
requests
waitress